[bug-reducer] Rename bug_reducer_utils.py => swift_tools.py
