+++ utils/bug_reducer/tests/test_optbugreducer.py
@@ -21,7 +21,7 @@ import subprocess
 import unittest
 
 
-import bug_reducer.bug_reducer_utils as bug_reducer_utils
+import bug_reducer.swift_tools as swift_tools
 
 
 @unittest.skipUnless(platform.system() == 'Darwin',
@@ -46,7 +46,7 @@ class OptBugReducerTestCase(unittest.TestCase):
         self.sdk = subprocess.check_output(['xcrun', '--sdk', 'macosx',
                                             '--toolchain', 'Default',
                                             '--show-sdk-path']).strip("\n")
-        self.tools = bug_reducer_utils.SwiftTools(self.build_dir)
+        self.tools = swift_tools.SwiftTools(self.build_dir)
         json_data = json.loads(subprocess.check_output(
             [self.tools.sil_passpipeline_dumper, '-Performance']))
         self.passes = []
